requirements.txt
streamlit
google-api-python-client
requests
python-dotenv
transformers
torch
nltk